release: 다중 계층 캐시 도입 및 시스템 성능/안정성 강화 (v2.0.0) (#87)
* feat: 장비 정보 계층형 캐시 도입 및 성능 최적화 (#83)

동일 캐릭터 반복 요청 시 발생하는 DB I/O와 JSON 파싱 비용을 줄이기
위해 Caffeine(L1) 및 DB AOP(L2)를 결합한 계층형 캐시를 구축함.
테스트 환경의 독립성을 보장하기 위해 각 테스트 시작 전 캐시를
초기화하는 로직과 비동기 모킹 방식을 개선함.

Issues #11

* refactor: CalculationCache 무한 성장 방지 및 정책 최적화 (#84)

큐브 시도 횟수 계산 결과의 무한 성장을 방지하기 위해 최대 크기를
10,000개로 제한함. 데이터 성격에 따라 TTL을 15분, 20분, 60분으로
차등 적용하여 메모리 효율과 데이터 정합성을 동시에 확보함.
중복된 캐시 생성 로직을 buildCaffeineCache 메서드로 단일화함.

Issues #12

* refactor: 계층형 좋아요 동기화 아키텍처 도입 및 장애 내성 강화 (#85)

서버 증설(Scale-out) 시 데이터 정합성을 보장하기 위해 Caffeine(L1)과
Redis(L2)를 결합한 계층형 버퍼 구조를 도입함. Redis 장애 시 DB로 직접
동기화하는 Fallback 로직을 추가하여 시스템 가용성을 극대화함.
가독성 개선을 위해 복잡한 람다 로직을 프라이빗 메서드로 분리함.

Issues #27

* feat: 계층형 캐시 도입 및 Redis 안정성 강화 (#86)

- Caffeine(L1)과 Redis(L2)를 조합한 다중 계층 캐시 구조 구현
- Jackson 역직렬화 시 Integer 캐스팅 에러 방지를 위해 반환 타입을 Double로 변경
- Docker-compose 내 Redis maxmemory 제한 및 LRU 방출 정책 설정 추가
- AOP 프록시 환경에서의 동시성 테스트 코드 정상화

+++ src/main/java/maple/expectation/service/v2/calculator/impl/BlackCubeDecorator.java
@@ -26,7 +26,7 @@ public class BlackCubeDecorator extends EnhanceDecorator {
     }
 
     public Long calculateTrials() {
-        trials = trialsProvider.calculateExpectedTrials(input, CubeType.BLACK);
+        trials = trialsProvider.calculateExpectedTrials(input, CubeType.BLACK).longValue();
         return trials;
     }
